% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_prisma.R
\name{convert_prisma}
\alias{convert_prisma}
\title{convert_prisma}
\usage{
convert_prisma(
  in_file,
  out_folder,
  out_filebase = "auto",
  out_format = "ENVI",
  base_georef = TRUE,
  fill_gaps = TRUE,
  VNIR = TRUE,
  SWIR = FALSE,
  FULL = FALSE,
  source = "HCO",
  join_priority = "SWIR",
  ATCOR = FALSE,
  ATCOR_wls = NULL,
  PAN = FALSE,
  CLOUD = FALSE,
  LC = FALSE,
  GLINT = FALSE,
  ANGLES = FALSE,
  LATLON = FALSE,
  ERR_MATRIX = FALSE,
  apply_errmatrix = FALSE,
  overwrite = FALSE,
  in_L2_file = NULL,
  selbands_vnir = NULL,
  selbands_swir = NULL
)
}
\arguments{
\item{in_file}{`character` full path of input HDF5 file}

\item{out_folder}{`character` full path to output folder}

\item{out_filebase}{`character` if "auto", output file names are based on
filename of the input prisma file, with appropriate suffixes. If a different
string is provided, it is used as prefix instead of the input file name,
Default: "auto"}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}

\item{base_georef}{`logical` if TRUE, apply base georeferencing on L1, L2B/C data,
using the "Georeference from input GLT" procedure explained here:
https://www.harrisgeospatial.com/docs/backgroundgltbowtiecorrection.html,
Default: TRUE}

\item{fill_gaps}{`logical` if TRUE, when georeferencing on L1, L2B/C data,
substitute missing values with results of a 3x3 focal filter on the georeferenced
data, Default: TRUE}

\item{VNIR}{`logical` if TRUE, create the VNIR image, Default: TRUE}

\item{SWIR}{`logical` if TRUE, create the SWIR image, Default: TRUE}

\item{FULL}{`logical` if TRUE, create a single multispectral image from
VNIR and SWIR, Default: FALSE}

\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HCO'}

\item{join_priority}{`character` ["VNIR" | "SWIR"], spectrometer to consider in
the when join_spectra = TRUE, Default: SWIR - ignored if join_spectra is FALSE.
Default: "VNIR"}

\item{ATCOR}{logical` if TRUE, create the text files required to run ATCOR, Default: FALSE;}

\item{ATCOR_wls}{NULL, or `numeric` If NULL the only ATCOR wvl file
created is the one containing Nominal wavelengths. If a numeric vector is provided,
then one different wvl file is created for each selected COLUMN selected (e.g., if
providing `ATCOR_wls = c(200, 800)`, then the wavelengths and FWHMs related to
columns 200 and 800 are saved.)}

\item{PAN}{`logical` if TRUE, also save PAN data, default: TRUE (Ignored for L2 data)}

\item{CLOUD}{`logical` if TRUE, also save CLOUD MASK mask data, default: TRUE (Ignored for L2 data)}

\item{LC}{`logical` if TRUE, also save the LAND COVER data, default: TRUE (Ignored for L2 data)}

\item{GLINT}{`logical` if TRUE, also save GLINT mask data, default: TRUE (Ignored for L2 data)}

\item{ANGLES}{if TRUE, also save the ACQUISITION ANGLES data. ANGLES data are saved as
a 3-band raster. Band 1 contains "obs_ang", Band 2 contains "relaz_ang" and
Band 3 contains "solzen_ang"), default: FALSE}

\item{LATLON}{if TRUE, also save the LATITUDE and LONGITUDE data. LATLON data are saved as
a 2-band raster. Band 1 contains "Lat", Band 2 contains "Lon", default: FALSE}

\item{ERR_MATRIX}{Not yet implemented!}

\item{apply_errmatrix}{Not yet implemented!}

\item{overwrite}{`logical` if TRUE, existing files are overwritten, default: FALSE}

\item{in_L2_file}{`character` full path of an L2 file to be used to extract better georeeferencing
info and angles for a corresponding L1 file. If not NULL, and `in_file` is a L1 file, the LAT and LON
fields used for bowtie georeferencing are taken from the L2 file instead than from the L1 file}

\item{selbands_vnir}{`numeric array` containing wavelengths (in nanometers) of bands that should be extracted from
the VNIR data cube. If not NULL, only the bands with wavelengths closest to these values are extracted, Default: NULL}

\item{selbands_swir}{`numeric array` containing wavelengths (in nanometers) of bands that should be extracted from
the SWIR data cube. If not NULL, only the bands with wavelengths closest to these values are extracted, Default: NULL}
}
\value{
The function is called for its side effects
}
\description{
Access a PRISMA L1 HDF5 file and convert it to ENVI or GeoTiff
 format
}
\examples{
\dontrun{
 in_file    <- "/home/lb/tmp/test/PRS_L1_STD_OFFL_20190825103112_20190825103117_0001.he5"
 out_folder <- "/home/lb/tmp/test/"
 out_format <- "ENVI"

 # Save VNIR Cube image
 convert_prisma(in_file       = in_file,
                out_folder    = out_folder,
                out_format    = out_format,
                VNIR          = TRUE,
                SWiR          = TRUE,
                FULL          = FALSE,
                overwrite     = TRUE
                )

 # Save also the full cube, prioritizing the VNIR spectrometer and save in ENVI format
 # Also save Cloud and LC data. Also, use a custom prefix for output files
 convert_prisma(in_file       = in_file,
                out_folder    = out_folder,
                out_filebase  = "prismaout_2020_05-14",
                out_format    = out_format,
                FULL          = TRUE,
                join_priority = "VNIR",
                LC            = TRUE,
                CLOUD         = TRUE,
                overwrite     = TRUE
                )

 # Save a full image, prioritizing the SWIR spectrometer and save in TIF format,
 # Also create ATCOR files, with Nominal wavelengths

 convert_prisma(in_file       = in_file,
                out_folder    = out_folder,
                out_format    = "TIF",
                FULL          = TRUE,
                join_priority = "VNIR",
                ATCOR         = TRUE,
                LC            = TRUE,
                CLOUD         = TRUE,
                overwrite     = TRUE
                )


}
}
