% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_addindex.R
\name{pr_addindex}
\alias{pr_addindex}
\title{pr_addindex}
\usage{
pr_addindex(Name, Formula, Description = NA, Sensitivity = NA, Reference = NA)
}
\arguments{
\item{Name}{`character` Name of the new index (e.g., "myindex")}

\item{Formula}{`character` Formula of the index. In the formula, bands
are to be referred to the prefix "R", followed by the required wavelength
/e.g., "R600 / R720"}

\item{Description}{`character`, Optional description of the index (e.g.,
"My Index for Nitrogen"), Default: NULL}

\item{Sensitivity}{`character`, Optional specification of index sensitivity
(e.g., "Nitrogen concentration"),  Default: NULL}

\item{Reference}{`character`, Optional specification of index literature
reference (e.g., "me et al., (2020)), Default: NULL}
}
\value{
NULL - The function is called for its side effects
}
\description{
Function used to add a spectral index to the list of available
o ones
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
pr_addindex(Name = "myindex", Formula = "R600 / R700",
           Description = "My custom Index",
           Reference = "Me (2020)")
 }
}
}
