% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_basegeo.R
\name{pr_basegeo}
\alias{pr_basegeo}
\title{pr_basegeo}
\usage{
pr_basegeo(band, lon, lat, fill_gaps = TRUE)
}
\arguments{
\item{band}{`Raster` layer to be georeferenced}

\item{lon}{`Matrix` containing the longitudes of each pixel (as derived from the
geolocation fields datasets of the hdf5)}

\item{lat}{`Matrix` containing the latitudes of each pixel (as derived from the
geolocation fields datasets of the hdf5)}

\item{fill_gaps}{`logical` If TRUE, pixels with no values in the georeferenced image are
filled based on a 3x3 average focal filter of the neighbouring valid pixels, Default: TRUE}
}
\value{
`Raster` georeferenced dataset
}
\description{
Function used to create a georeferenced raster from L1/L2B/L2C datasets,
 based on the Latitude and Longitude data, and exploiting the procedure described in
 https://www.harrisgeospatial.com/docs/backgroundgltbowtiecorrection.html
}
\details{
The function is based on the "GLT and Bowtie Correction" technique used in ENVI, and
described in https://www.harrisgeospatial.com/docs/backgroundgltbowtiecorrection.html.
Note that the 7x7 interpolation step for pixels still missing after the 3x3 interpolation is not (yet)
implemented, as well as the nearest neighbour interpolation.
}
\seealso{
\code{\link[raster]{getValues}}
}
