% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_check_formulas.R
\name{pr_check_formulas}
\alias{pr_check_formulas}
\title{pr_check_formulas}
\usage{
pr_check_formulas(indexes)
}
\arguments{
\item{indexes}{either a named list containing indexes names and formulas (
e.g., indexes = list(myindex1 = "R500 / R600",
                     myindex2 = "(R800 - R680) / (R800 + R680)")), or a
data.frame containing the columns "Name" and "Formula".}
}
\value{
NULL the function is called for its side effects
}
\description{
Helper function used to check if formula of a spectral index
 can be computed
}
\details{
DETAILS
}
\examples{
\dontrun{
 pr_check_formulas(indexes = list(myindex1 = "R500 / R600",
                      myindex2 = "(R800 - R680) / (R800 + R680)"))
}
}
