% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_listindexes.R
\name{pr_listindexes}
\alias{pr_listindexes}
\title{pr_listindexes}
\usage{
pr_listindexes()
}
\value{
NULL - The function is called for its side effects
}
\description{
helper function used to show the user a DT containing the list
 of available spectral indexes
}
