% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_make_atcor.R
\name{pr_make_atcor}
\alias{pr_make_atcor}
\title{pr_make_atcor}
\usage{
pr_make_atcor(
  f,
  out_file,
  ATCOR_wls,
  wls,
  fwhms,
  order_vnir,
  order_swir,
  join_priority,
  source
)
}
\arguments{
\item{f}{input data he5 from caller}

\item{out_file}{main prismaread output file name - passed by caller}

\item{ATCOR_wls}{NULL, or `numeric` If NULL the only ATCOR wvl file
created is the one containing Nominal wavelengths. If a numeric vector is
provided, then one different wvl file is created for each selected COLUMN
selected (e.g., if providing `ATCOR_wls = c(200, 800)`, then the wavelengths
and FWHMs related to columns 200 and 800 are saved.)}

\item{wls}{passed by caller - array of PRISMA wavelengths}

\item{fwhms}{passed by caller - array of PRISMA fwhms}

\item{order_vnir}{passed by caller - ordering of array of PRISMA VNIR
wavelengths}

\item{order_swir}{passed by caller - ordering of array of PRISMA SWIR
wavelengths}

\item{join_priority}{`character` ["VNIR" | "SWIR"], spectrometer to consider
when join_spectra = TRUE, Default: SWIR - ignored if join_spectra is FALSE.
Default: "SWIR"}

\item{source}{`character` ["HC0" | "HRC"], Considered Data Cub,
Default: 'HCO'}
}
\value{
`list` containing required info according to `proc_lev`
}
\description{
helper function used to create accessory files needed for ATCOR
 atmospheric correction
}
