% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_rastwrite_lines.R
\name{pr_rastwrite_lines}
\alias{pr_rastwrite_lines}
\title{pr_rastwrite_lines}
\usage{
pr_rastwrite_lines(
  rast_in,
  out_file,
  out_format = "tif",
  proc_lev = "1",
  scale_min = NULL,
  scale_max = NULL,
  join = FALSE
)
}
\arguments{
\item{rast_in}{`Raster* object` to be written to disk}

\item{out_file}{`character` full path of output image}

\item{out_format}{`character` [\"TIF\" | \"ENVI\"], Default: 'tif'}

\item{proc_lev}{`character` [\"1\" | \"2D\"], Default: '1'}

\item{scale_min}{`numeric` coefficients use to compute values from DN on
products}

\item{scale_max}{`numeric`  coefficients use to compute values from DN on L2
products}

\item{join}{`logical` flag used to indicate if we are saving the "joined"
VNIR+SWIR cube}
}
\value{
the function is called for its side effects
}
\description{
Write a raster object by blocks of lines
}
\details{
DETAILS
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
