% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_basegeo.R
\name{prisma_basegeo}
\alias{prisma_basegeo}
\title{FUNCTION_TITLE}
\usage{
prisma_basegeo(band, lon, lat, fill_gaps = TRUE)
}
\arguments{
\item{band}{PARAM_DESCRIPTION}

\item{lon}{PARAM_DESCRIPTION}

\item{lat}{PARAM_DESCRIPTION}

\item{fill_gaps}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[raster]{getValues}}
}
