% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_cld.R
\name{prisma_create_cld}
\alias{prisma_create_cld}
\title{prisma_create_cld}
\usage{
prisma_create_cld(f, out_file_cld, out_format, base_georef)
}
\arguments{
\item{f}{input data he5 from caller}

\item{out_file_cld}{output file name for glint}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}

\item{base_georef}{`logical` if TRUE, apply base georeferencing on L1, L2B/C data,
Default: FALSE}
}
\value{
The function is called for its side effects
}
\description{
helper function used to process and save the CLOUD data cube
}
