% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_pan.R
\name{prisma_create_pan}
\alias{prisma_create_pan}
\title{prisma_create_pan}
\usage{
prisma_create_pan(f, proc_lev, source, out_file_pan, out_format, base_georef)
}
\arguments{
\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
helper function used to process and save the VNIR data cube
}
\details{
DETAILS
}
