% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_swir.R
\name{prisma_create_swir}
\alias{prisma_create_swir}
\title{prisma_create_swir}
\usage{
prisma_create_swir(
  f,
  proc_lev,
  source,
  out_file_swir,
  out_format,
  base_georef,
  wl_swir,
  order_swir,
  fwhm_swir,
  apply_errmatrix,
  ERR_MATRIX
)
}
\arguments{
\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}

\item{wl_swir}{passed by caller - array of PRISMA SWIR wavelengths}

\item{order_swir}{passed by caller - ordering of array of PRISMA VNIR wavelengths}

\item{fwhm_swir}{passed by caller - array of PRISMA VNIR fwhms}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
