% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_vnir.R
\name{prisma_create_vnir}
\alias{prisma_create_vnir}
\title{prisma_create_vnir}
\usage{
prisma_create_vnir(
  f,
  proc_lev,
  source,
  out_file_vnir,
  out_format,
  base_georef,
  wl_vnir,
  order_vnir,
  fwhm_vnir,
  apply_errmatrix,
  ERR_MATRIX
)
}
\arguments{
\item{f}{input data he5 from caller}

\item{proc_lev}{`character` Processing level (e.g., "1", "2B") - passed by caller}

\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'}

\item{out_file_vnir}{output file name for VNIR}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}

\item{base_georef}{`logical` if TRUE, apply base georeferencing on L1, L2B/C data,
Default: FALSE}

\item{wl_vnir}{passed by caller - array of PRISMA VNIR wavelengths}

\item{order_vnir}{passed by caller - ordering of array of PRISMA VNIR wavelengths}

\item{fwhm_vnir}{passed by caller - array of PRISMA VNIR fwhms}

\item{apply_errmatrix}{Not yet implemented!}

\item{ERR_MATRIX}{Not yet implemented!}
}
\value{
the function is called for its side effects
}
\description{
helper function used to process and save the VNIR data cube
}
