% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastwrite_lines.R
\name{rastwrite_lines}
\alias{rastwrite_lines}
\title{rastwrite_lines}
\usage{
rastwrite_lines(rast_in, out_file, out_format = "tif")
}
\arguments{
\item{rast_in}{`Raster* object` to be written to disk}

\item{out_file}{`character` full path of output image}

\item{out_format}{`character` [\"TIF\" | \"ENVI\"], Default: 'tif'}
}
\value{
the function is called for its side effects
}
\description{
Write a raster object by blocks of lines
}
\details{
DETAILS
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
