% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_prisma.R
\name{convert_prisma}
\alias{convert_prisma}
\title{convert_prisma}
\usage{
convert_prisma(in_file, out_file, out_format = "ENVI", VNIR = TRUE,
  SWIR = TRUE, FULL = FALSE, source = "HRC",
  join_priority = "SWIR", PAN = TRUE, CLOUD = FALSE, LC = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{in_file}{`character` full path of input HDF5 file}

\item{out_file}{`character` full path of output  file}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}

\item{VNIR}{`logical` if TRUE, create the VNIR image, Default: TRUE}

\item{SWIR}{`logical` if TRUE, create the SWIR image, Default: TRUE}

\item{FULL}{`logical` if TRUE, create a single multispectral image from
VNIR and SWIR, Default: FALSE}

\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'}

\item{join_priority}{`character` ["VNIR" | "SWIR"], spectrometer to consider in
the when join_spectra = TRUE, Default: SWIR - ignored if join_spectra is FALSE}

\item{PAN}{`logical` if TRUE, also save the PAN data, default: TRUE}

\item{CLOUD}{`logical` if TRUE, also save the cloud mask data, default: TRUE}

\item{LC}{`logical` if TRUE, also save the land cover data, default: TRUE}

\item{overwrite}{`logical` if TRUE, existing files are overwritten, default: FALSE}
}
\value{
The function is called for its side effects
}
\description{
Access a PRISMA L1 HDF5 file and convert it to ENVI or GeoTiff
 format
}
\examples{
\dontrun{
 in_file  <- "/home/lb/tmp/test/PRS_L1_STD_OFFL_20190825103112_20190825103117_0001.he5"
 out_file <- "/home/lb/tmp/test/test_1"
 out_format <- "ENVI"


 # Save a full image, prioritizing the SWIR spectrometer and save in EVI format
 convert_prisma(in_file       = in_file,
                out_file      = out_file,
                out_format    = out_format,
                FULL          = TRUE,
                join_priority = "SWIR",
                LC            = TRUE,
                CLOUD         = TRUE
                )
}
}
