% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_prisma.R
\name{convert_prisma}
\alias{convert_prisma}
\title{convert_prisma}
\usage{
convert_prisma(
  in_file,
  out_file,
  out_format = "ENVI",
  base_georef = FALSE,
  VNIR = TRUE,
  SWIR = TRUE,
  FULL = FALSE,
  source = "HCO",
  join_priority = "SWIR",
  ATCOR = FALSE,
  ATCOR_wls = NULL,
  PAN = TRUE,
  CLOUD = FALSE,
  LC = FALSE,
  GLINT = FALSE,
  ERR_MATRIX = FALSE,
  apply_errmatrix = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{in_file}{`character` full path of input HDF5 file}

\item{out_file}{`character` full path of output  file}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}

\item{base_georef}{`logical` if TRUE, apply base georeferencing on L1, L2B/C data,
Default: FALSE}

\item{VNIR}{`logical` if TRUE, create the VNIR image, Default: TRUE}

\item{SWIR}{`logical` if TRUE, create the SWIR image, Default: TRUE}

\item{FULL}{`logical` if TRUE, create a single multispectral image from
VNIR and SWIR, Default: FALSE}

\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'}

\item{join_priority}{`character` ["VNIR" | "SWIR"], spectrometer to consider in
the when join_spectra = TRUE, Default: SWIR - ignored if join_spectra is FALSE.
Default: "VNIR"}

\item{ATCOR}{logical` if TRUE, create the text files required to run ATCOR, Default: FALSE;}

\item{ATCOR_wls}{NULL, or `numeric` If NULL the only ATCOR wvl file
created is the one containing Nominal wavelengths. If a numeric vector is provided,
then one different wvl file is created for each selected COLUMN selected (e.g., if
providing `ATCOR_wls = c(200, 800)`, then the wavelengths and FWHMs related to
columns 200 and 800 are saved.)}

\item{PAN}{`logical` if TRUE, also save the PAN data, default: TRUE}

\item{CLOUD}{`logical` if TRUE, also save the cloud mask data, default: TRUE}

\item{LC}{`logical` if TRUE, also save the land cover data, default: TRUE}

\item{GLINT}{`logical` if TRUE, also save the GLINT mask data, default: TRUE}

\item{ERR_MATRIX}{Not yet implemented!}

\item{apply_errmatrix}{Not yet implemented!}

\item{overwrite}{`logical` if TRUE, existing files are overwritten, default: FALSE}
}
\value{
The function is called for its side effects
}
\description{
Access a PRISMA L1 HDF5 file and convert it to ENVI or GeoTiff
 format
}
\examples{
\dontrun{
 in_file  <- "/home/lb/tmp/test/PRS_L1_STD_OFFL_20190825103112_20190825103117_0001.he5"
 out_file <- "/home/lb/tmp/test/test_1"
 out_format <- "ENVI"

 # Save VNIR Cube image
 convert_prisma(in_file       = in_file,
                out_file      = out_file,
                out_format    = out_format,
                VNIR          = TRUE,
                SWiR          = TRUE,
                FULL          = FALSE,
                overwrite     = TRUE
                )

 # Save also the full cube, prioritizing the VNIR spectrometer and save in ENVI format
 # Also save Cloud and LC data
 convert_prisma(in_file       = in_file,
                out_file      = out_file,
                out_format    = out_format,
                FULL          = TRUE,
                join_priority = "VNIR",
                LC            = TRUE,
                CLOUD         = TRUE,
                overwrite     = TRUE
                )

 # Save a full image, prioritizing the SWIR spectrometer and save in ENVI format,
 # Also create ATCOR files, with Nominal wavelengths, and those for columns
 # 200 and 800 of the cube
 convert_prisma(in_file       = in_file,
                out_file      = out_file,
                out_format    = out_format,
                FULL          = TRUE,
                join_priority = "VNIR",
                ATCOR         = TRUE,
                ATCOR_wls     = c(200,800),
                LC            = TRUE,
                CLOUD         = TRUE,
                overwrite     = TRUE
                )


}
}
