% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_cld.R
\name{prisma_create_cld}
\alias{prisma_create_cld}
\title{prisma_create_cld}
\usage{
prisma_create_cld(f, out_file_cld, out_format, base_georef)
}
\arguments{
\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
helper function used to process and save the VNIR data cube
}
\details{
DETAILS
}
