% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_glint.R
\name{prisma_create_glint}
\alias{prisma_create_glint}
\title{prisma_create_glint}
\usage{
prisma_create_glint(f, out_file_glint, out_format, base_georef)
}
\arguments{
\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}
}
\value{
The function is called for its side effects
}
\description{
helper function used to process and save the GLINT data cube
}
