% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_lc.R
\name{prisma_create_lc}
\alias{prisma_create_lc}
\title{prisma_create_lc}
\usage{
prisma_create_lc(f, out_file_lc, out_format, base_georef)
}
\arguments{
\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}
}
\value{
The function is called for its side effects
}
\description{
helper function used to process and save the LC data cube
}
