% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_swir.R
\name{prisma_create_swir}
\alias{prisma_create_swir}
\title{prisma_create_swir}
\usage{
prisma_create_swir(
  f,
  proc_lev,
  source,
  out_file_swir,
  out_format,
  base_georef,
  wl_swir,
  order_swir,
  fwhm_swir,
  apply_errmatrix,
  ERR_MATRIX
)
}
\arguments{
\item{f}{input data he5 from caller}

\item{proc_lev}{`character` Processing level (e.g., "1", "2B") - passed by caller}

\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'}

\item{out_file_swir}{output file name for SWIR}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}

\item{base_georef}{`logical` if TRUE, apply base georeferencing on L1, L2B/C data,
Default: FALSE}

\item{wl_swir}{passed by caller - array of PRISMA SWIR wavelengths}

\item{order_swir}{passed by caller - ordering of array of PRISMA SWIR wavelengths}

\item{fwhm_swir}{passed by caller - array of PRISMA SWIR fwhms}

\item{apply_errmatrix}{Not yet implemented!}

\item{ERR_MATRIX}{Not yet implemented!}
}
\value{
the function is called for its side effects
}
\description{
helper function used to process and save the SWIR data cube
}
