% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_create_vnir.R
\name{prisma_create_vnir}
\alias{prisma_create_vnir}
\title{prisma_create_vnir}
\usage{
prisma_create_vnir(
  f,
  proc_lev,
  source,
  out_file_vnir,
  out_format,
  base_georef,
  wl_vnir,
  order_vnir,
  fwhm_vnir,
  apply_errmatrix,
  ERR_MATRIX,
  proj_code
)
}
\arguments{
\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'}

\item{out_format}{`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'}

\item{wl_vnir}{PARAM_DESCRIPTION}

\item{order_vnir}{PARAM_DESCRIPTION}

\item{fwhm_vnir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
helper function used to process and save the VNIR data cube
}
\details{
DETAILS
}
