% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisma_get_geoloc.R
\name{prisma_get_geoloc}
\alias{prisma_get_geoloc}
\title{prisma_get_geoloc}
\usage{
prisma_get_geoloc(f, proc_lev, source, wvl = NULL)
}
\arguments{
\item{f}{input data he5 from caller}

\item{proc_lev}{`character` Processing level (e.g., "1", "2B") - passed by caller}

\item{source}{`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'}
}
\value{
`list` containing required info according to `proc_lev`
}
\description{
helper function used to get geolocation info
 from PRISMA data on VNIR and SWIR cubes
}
